github_pr_monitor/requirement.txt
rumps==0.4.0
keyring==24.3.0
PyGithub==2.1.1
